# code/Java_Projects/M2-Challenge-Schmidt-Jeff/openapi.yaml
openapi: 3.0.1
info:
  title: Module 2 Challenge
  description: "This is API documentation for the Month and Math API I created, with a lot of help from Office Hours and Class Activities."
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /subtract:
    post:
      tags:
        - math-solution-controller
      operationId: subtractNumbers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /multiply:
    post:
      tags:
        - math-solution-controller
      operationId: multiplyNumbers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /divide:
    post:
      tags:
        - math-solution-controller
      operationId: divideNumbers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /add:
    post:
      tags:
        - math-solution-controller
      operationId: addNumbers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /randomMonth:
    get:
      tags:
        - month-controller
      operationId: getRandomMonth
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /month:
    get:
      tags:
        - month-controller
      operationId: getAllMonths
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
  /month/{number}:
    get:
      tags:
        - month-controller
      operationId: getMonthById
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
                  - $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    MathSolution:
      type: object
      properties:
        operand1:
          type: integer
          format: int32
        operand2:
          type: integer
          format: int32
        answer:
          type: integer
          format: int32
        operations:
          type: string
    Month:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
          format: int32